Remove HMAC and tweak, as they added unnecessary complexity
